+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 * Unreleased
+    * If there are duplicate normalized zone names or link names, throw an
+      exception to make it a fatal condition.
+        * Allows detection and fixing of this problem, instead of silently
+          dropping zones or links.
 * v1.1.4 (2022-11-02)
     * Add `--generate_int16_year` flag to generate year fields with `int16_t`
       type instead of `int8_t`.